+++ CHANGELOG.md
@@ -30,6 +30,7 @@ Not yet released
 * Fix rendering of literals in JPA Criteria and introduce configuration option to control value rendering
 * Fix base URI determination in Spring HATEOAS integration to be compatible with Spring HATEOAS 1.3.4
 * Make sure `LIMIT`/`OFFSET` is respected when generating a count query
+* Fix concurrency issue in entity view annotation processor leading to strange errors
 
 ### Backwards-incompatible changes